pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "litscout"
version = "0.1.0"
description = "Scientific literature search, citation expansion, and full-text retrieval toolkit"
requires-python = ">=3.11"
license = "MIT"
readme = "README.md"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "click>=8.0",
    "semanticscholar>=0.8",
    "pyalex>=0.14",
    "biopython>=1.84",
    "pymupdf4llm>=0.0.17",
    "pubmed_parser>=0.4",
    "rapidfuzz>=3.0",
    "tqdm>=4.66",
    "tenacity>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff",
]

[project.scripts]
litscout = "litscout.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
